Sync with upstream fork

+++ tests/nvme_create_max_ns_test.py
@@ -50,7 +50,7 @@ class TestNVMeCreateMaxNS(TestNVMe):
         TestNVMe.__init__(self)
         self.dps = 0
         self.flbas = 0
-        self.nsze = self.get_ncap() / self.get_format() / self.get_max_ns()
+        self.nsze = int(self.get_ncap() / self.get_format() / self.get_max_ns())
         self.ncap = self.nsze
         self.setup_log_dir(self.__class__.__name__)
         self.max_ns = self.get_max_ns()
